Revert "Revert "fixing database and google sheet""
