+++ backend/localizations/translations/de-DE.json
@@ -3,20 +3,21 @@
     "drop-down-label-en-US": "Englisch",
     "drop-down-label-fr-FR": "Französisch",
     "drop-down-label-ko-KR": "Koreanisch",
+    "drop-down-label-de-DE":"Deutsh",
     "page-title": "Willkommen bei Twiri",
     "logoText": "Twiri, finde deinen Stream",
     "copyright-text": "Copyright © 2020 JinriTV",
-    "animated-words-label": "Du weißt nicht wen du..",
-    "animated-words": ["anschauen sollst?", "hosten sollst?", "folgen sollst?", "raiden sollst?"],
+    "animated-words-label": "Du weißt nicht wen <br> du ",
+    "animated-words": ["anschauen sollst", "hosten sollst", "folgen sollst", "raiden sollst"],
     "find-streamer-button": "Finde deinen Streamer",
-    "question-text-languages": "Welche Sprachen?",
+    "question-text-languages": "Welche Sprachen hörst du dir gerne an?",
     "button-text-languages-english": "Englisch",
     "button-text-languages-korean":"Koreanisch",
     "button-text-languages-japanese": "Japanisch",
     "button-text-languages-chinese": "Chinesisch",
     "button-text-languages-french": "Französisch",
     "button-text-languages-spanish": "Spanisch",
-    "question-text-content":  "Ich bevorzuge diese Streamer...",
+    "question-text-content":  "Welchen Content bevorzugst du?",
     "button-text-content-justchatting": "Just Chatting",
     "button-text-content-games": "Games",
     "button-text-content-ASMR": "ASMR",
@@ -29,14 +30,14 @@
     "button-text-content-music": "Musik",
     "button-text-content-dancing": "Tanzen",
     "button-text-content-yoga": "Yoga",
-    "question-text-subonly": "Wer kann chatten?",
+    "question-text-subonly": "Wer soll chatten können?",
     "button-text-subonly-all":"Jeder",
     "button-text-subonly-sub-only": "Nur Subs",
     "button-text-subonly-follower-only": "Nur Follower",
-    "question-text-mature": "Welcher Content?",
+    "question-text-mature": "Bevorzugst du Kanäle für Erwachsene oder Familien?",
     "button-text-mature-true": "Inhalte für Erwachsene",
     "button-text-mature-false": "Familienfreundlich",
-    "question-text-chat-vibe": "Stimmung des Chats",
+    "question-text-chat-vibe": "Welche Stimmung sollte der Chat herüberbringen?",
     "button-text-chat-vibe-chatty": "Gesprächig",
     "button-text-chat-vibe-chill": "Chillig",
     "button-text-chat-vibe-serious": "Seriös",
@@ -60,8 +61,8 @@
     "button-text-chat-vibe-slow": "Langsamer Chat",
     "button-text-chat-vibe-wholesome": "Verträglich",
     "button-text-chat-vibe-toxic": "Toxic",
-    "question-text-average_viewers": "Durchschnittliche Zuschaueranzahl",
-    "question-text-watchtime": "An diesen Tagen kann ich zuschauen..",
+    "question-text-average_viewers": "Mit welcher Zuschaueranzahl bist du zufrieden?",
+    "question-text-watchtime": "An welchen Tagen/Uhrzeiten willst du Streams anschauen?",
     "time-range-watchtime-weekdays": "Wochentagen",
     "time-range-watchtime-weekends": "Wochenenden",
     "time-range-watchtime-from": "von",
