+++ backend/localizations/translations/en-US.json
@@ -1,13 +1,15 @@
 {   "this-language": "en-US",
     "dark-mode-label": "dark mode?",
+    "light-mode-label":"light mode?",
     "drop-down-label-en-US": "English",
     "drop-down-label-fr-FR": "French",
     "drop-down-label-ko-KR": "Korean",
+    "drop-down-label-de-DE":"German",
     "page-title": "Welcome to Twiri",
     "logoText": "Twiri, Stream seeker",
     "copyright-text": "Copyright © 2020 JinriTV",
-    "animated-words-label": "Don't know who to",
-    "animated-words": ["watch?", "host?", "follow?", "raid?"],
+    "animated-words-label": "Don't know who <br> to ",
+    "animated-words": ["watch", "host", "follow", "raid"],
     "find-streamer-button": "Find a Streamer",
     "question-text-languages": "Spoken languages?",
     "button-text-languages-english": "English",
