+++ backend/localizations/translations/fr-FR.json
@@ -3,11 +3,12 @@
     "drop-down-label-en-US": "Anglais",
     "drop-down-label-fr-FR": "Français",
     "drop-down-label-ko-KR": "Coréen",
+    "drop-down-label-de-DE":"Allemagne",
     "page-title": "Bienvenue à Twiri",
     "logoText": "Twiri, investigatrice des streams",
     "copyright-text": "Copyright © 2020 JinriTV",
-    "animated-words-label": "Qui",
-    "animated-words": ["regarder ?", "héberger ?", "suivre ?", "raider ?"],
+    "animated-words-label": "Qui ",
+    "animated-words": ["regarder", "héberger", "suivre", "raider"],
     "find-streamer-button": "Rechercher un streamer",
     "question-text-languages": "Dans quelle(s) langue(s) ?",
     "button-text-languages-english": "Anglais",
