+++ package.json
@@ -6,7 +6,7 @@
     "start": "nodemon app.js"
   },
   "dependencies": {
-    "axios": "^0.20.0",
+    "axios": "^0.21.1",
     "cls-hooked": "^4.2.2",
     "d3-dsv": "^2.0.0",
     "debug": "~2.6.9",
