+++ public/index.html
@@ -85,7 +85,7 @@
   <div class="footer d-flex justify-content-center">
     <!-- Copyright Section-->
     <div id="copyright-text"></div>
-    <div onclick="toggleDarkMode()" id="dark-mode-toggle"><span id="dark-mode-toggle-icon">🌚</span><span id="dark-mode-label"></span></div>
+    <div onclick="toggleDarkMode()" id="dark-mode-toggle"><span id="dark-mode-toggle-icon"></span><span id="dark-mode-label"></span></div>
   </div>
 
   <!-- Loading Scripts -->
