+++ public/new_index.html
@@ -1,0 +1,138 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta name="description" content="">
+  <meta name="author" content="">
+  <!-- Favicon-->
+  <link rel="icon" type="image/x-icon" href="/images/favicon.ico?v=1.1" />
+  <!-- Font Awesome icons (free version)-->
+  <script src="https://use.fontawesome.com/releases/v5.13.0/js/all.js" crossorigin="anonymous"></script>
+  <!-- Google fonts-->
+  <link rel="preconnect" href="https://fonts.gstatic.com">
+  <link href="https://fonts.googleapis.com/css2?family=Grandstander:wght@200&display=swap" rel="stylesheet">
+  <!-- Animations -->
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.0/animate.min.css"
+    integrity="sha256-6hqHMqXTVEds1R8HgKisLm3l/doneQs+rS1a5NLmwwo=" crossorigin="anonymous" />
+  <link href="stylesheets/animations.css" rel="stylesheet" />
+  <!-- Core theme CSS (includes Bootstrap) -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
+    integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+  <link href="stylesheets/bootstrap-slider.css" rel="stylesheet" />
+
+
+
+  <!-- Our custom CSS -->
+  <link href="stylesheets/quizstyle.css" rel="stylesheet" />
+  <link href="stylesheets/themes.css" rel="stylesheet" />
+  <link href="starry-rating/starry.min.css" rel="stylesheet" />
+</head>
+
+<body  id="page-top" class="light-mode">
+
+  <header>
+    <!-- Fixed navbar -->
+    <div class="page-header">
+      <div class="container d-flex justify-content-between" id="logo-container">
+
+        <a href="https://streamer-match-maker.herokuapp.com" class="navbar-brand d-flex align-items-center">
+          <img src="./images/logo.svg" style="width:45px;height:45px;margin-right:10px;" id="TwiriLogo" />
+          <span id="logoText"></span>
+        </a>
+
+        <a class="navbar-brand d-flex align-items-center">
+          <div id="generated-language-dropdown">
+          </div>
+        </a>
+      </div>
+
+    </div>
+  </header>
+
+  <!-- Begin page content -->
+  <main id="main-container" role="main" class="container">
+
+    <div id="welcome-banner" style="z-index:777;" class="welcome-container row">
+
+      <div id="welcome-text" style="z-index:777; opacity: 0;" class="col-lg-6">
+        <div class="row animated-words-container align-items-center">
+          <span><span id="animated-words-label" class="animated-words dark-text"></span><span class="animated-words"
+              id="keyword"></span></span>
+
+        </div>
+
+        <p class="info-text">
+          Twiri helps you discover new Twitch streamers <br>who are just your type.
+        </p>
+        <span class="matches-made-text"><span id="num-matches">2000</span>+ matches made</span>
+
+        <div class="row">
+          <button type="button" onclick="openGeneratedQuizModal()" class="btn btn-start-quiz"><span
+              id="find-streamer-button" style="font-size: 1.5em; color: cornsilk; "></span></button>
+
+        </div>
+      </div>
+      <div style="z-index:777;" class="col-lg-6 justify-content-center">
+        <div class="detective" style="opacity:0;">
+          <img class="img_detective" src="/images/mascot.png"/>
+        </div>
+      </div>
+    </div>
+    <div id="generated-result-page" style="z-index:500;" class="results-container">
+
+    </div>
+    <div class="background-container confetti-container">
+      <img src="./images/streamer1.png" class="streamer-image profile-1">
+      <img src="./images/streamer2.png" class="streamer-image profile-2">
+      <img src="./images/streamer3.png" class="streamer-image profile-3">
+      <img src="./images/streamer4.png" class="streamer-image profile-4">
+      <img src="./images/streamer5.png" class="streamer-image profile-5">
+      <img src="./images/streamer6.png" class="streamer-image profile-6">
+      <img src="./images/streamer7.png" class="streamer-image profile-7">
+      <img src="./images/streamer8.png" class="streamer-image profile-8">
+      <div id="sm-circle" class="base-circle sphere-small"></div>
+      <div id="md-circle" class="base-circle sphere-med"></div>
+      <div id="lg-circle" class="base-circle sphere-lg"></div>
+    </div>
+    </div>
+  </main>
+  <div class="page-footer">
+    <div class="container d-flex justify-content-between">
+      <span class="footer-text">About | Submit your stream | Contribute to the project</span>
+      <span class="footer-text">Copyright 2021</span>
+      
+  <div onclick="toggleDarkMode()" id="dark-mode-toggle"><span id="theme-icon"></span><span id="theme-label"></span></div>
+    
+    
+    </div>
+  </div>
+
+  <!-- Bootstrap core JavaScript
+    ================================================== -->
+    <!-- Bootstrap core JS-->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js"></script>
+  <!-- Bootstrap core JS-->
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js"></script>
+  <!-- Third party plugin JS-->
+
+  <!-- Core theme JS-->
+  <script src="javascripts/bootstrap-slider.min.js"></script>
+  <script src="javascripts/html_helpers.js"></script>
+  <script src="javascripts/create_quiz.js"></script>
+  <script src="javascripts/index.js"></script>
+  <script src="javascripts/jquery-clock-timepicker.min.js"></script>
+  <script src="starry-rating/starry.min.js"></script>
+</body>
+
+<!-- Generated Modal (questions are added here after page loads) -->
+<div class="modal fade" id="generated-quiz-modal" tabindex="-1" role="dialog">
+  <!---
+        the question pages are created and added in here after the page loads. Check the BuildQuiz() function
+        in the create_quiz.js script file.
+      -->
+</div>
+
+</html>
