+++ public/stylesheets/quizstyle.css
@@ -2,10 +2,14 @@
 * {
     font-family: 'Grandstander', cursive;
 }
-
 #dark-mode-toggle{
     cursor: pointer;
 }
+
+.wide-container{
+    max-width:90vw;
+}
+
 
 .dropdown-lang-icon {
     max-width: 34px;
@@ -27,28 +31,11 @@
     line-height: inherit;
     white-space: nowrap;
 }
-.btn {
-    display: inline-block;
-    font-weight: 400;
-    text-align: center;
-    vertical-align: middle;
-    cursor: pointer;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    background-color: transparent;
-    border-width: .125rem;
-    border-style: solid;
-    padding: .375rem .75rem;
-    font-size: 1rem;
-    line-height: 1.5;
-    border-radius: .5rem;
-    transition: color .15s ease-in-out,background-color .15s ease-in-out,border-color .15s ease-in-out,box-shadow .15s ease-in-out;
-}
+
 .btn.disabled, .btn:disabled {
     opacity: .65;
 }
+button { outline: none !important; box-shadow: none !important; }
 
 .modal-footer {
     display: flex;
@@ -104,11 +91,24 @@
     border-radius: .5rem;
 }
 
+.crown-container{
+    z-index: 555;
+    overflow: visible;
+    position: relative;
+}
+
+.crown{
+    position: absolute;
+    top: -96px;
+    right: -96px;
+}
+
+.socials-container{
+    margin-top: 48px;
+}
 
 .language-dropdown {
     width:100%;
-    border-width: 2px;
-    border-radius: 4px;
 }
 body {
     margin: 0;
@@ -118,8 +118,12 @@ body {
     text-align: left;
 }
 .dropdown-item-container{
+  
     padding-left:4px;
     padding-right:8px;
+}
+.dropdown-menu{
+    padding:0;
 }
 
 .dropdown-item {
@@ -135,7 +139,25 @@ body {
     border-radius: 24px; 
     border-width: 2px; 
     border-style: solid;
+    height:60vh;
+    text-align: center;
+    position: relative;
 }
+
+.percent-match{
+    font-size: xx-large;
+}
+
+.others{
+    margin-top:24px;
+    height: 55vh;
+}
+
+.winner{
+    border-color: goldenrod !important;
+    border-width: 4px; 
+}
+
 
 .modal-footer{
     position: absolute; 
@@ -183,15 +205,27 @@ body {
     margin: 24px;
 }
 
+.streamer-username{
+    font-size: xx-large;
+}
+
 .quiz-button-container{
-    padding-left:30%;
-    padding-right:30%;
+    padding-left:10%;
+    padding-right:10%;
 }
 
 .btn-quiz-continue {
     border-radius: 100px !important;
     margin-bottom: 12px;
     max-width:50%;
+}
+
+.btn-quiz-restart {
+    border-radius: 100px !important;
+    margin-bottom: 12px;
+    max-width:30%;
+    margin-left:auto;
+    margin-right:auto;
 }
 
 .btn-quiz-back{
@@ -242,9 +276,32 @@ body {
 }
 
 .welcome-container {
-    margin-top:150px;
+    margin-top:30vh;
+    margin-left:5vw;
 }
 
+.results-container{
+    margin-top:15vh;
+    position: relative;
+}
+
+.animated-words{
+    font-weight: bold;
+    font-size: 3em;
+}
+
+.animated-words-container{
+
+}
+
+
+.info-text{
+    font-size:1em;
+}
+
+.matches-made-text{
+    font-size: 2.5em;
+}
 
 .footer {   
     position:fixed;
@@ -270,10 +327,6 @@ body {
     min-height: 525px;
 }
 
-.modal-dialog {
-    max-width:40%;
-    margin:1.75rem auto
-   }
 
   .center {
     -webkit-box-align: center;
@@ -315,4 +368,310 @@ Unique elements (id)
 
 #dark-mode-toggle-icon{
     margin-right: 4px;
+}
+
+.blurred-profiles-container{
+    width:100%; 
+    height:100%;
+    position: relative;
+    z-index: 777;
+}
+
+.streamer-image {
+    /* Add the blur effect */
+    filter: blur(8px);
+    -webkit-filter: blur(8px);
+    border-radius: 50%;
+    /* Full height */
+    height: 100px;
+    width: 100px;
+    position: absolute;
+    z-index: 777;
+}
+  
+.profile-1{
+    height: 100px;
+    width: 100px;
+    top:20vh;
+	bottom: 0;
+	left: 10vw;
+	right: 0;
+}
+
+.profile-2{
+    height: 140px;
+    width: 140px;
+    top:75vh;
+	bottom: 0;
+	left: 10vw;
+	right: 0;
+}
+.profile-3{
+    height: 100px;
+    width: 100px;
+    top:70vh;
+	bottom: 0;
+	left: 15vw;
+	right: 0;
+}
+.profile-4{
+    height: 120px;
+    width: 120px;
+    top:25vh;
+	bottom: 0;
+	left: 55vw;
+	right: 0;
+}
+.profile-5{
+    height: 120px;
+    width: 120px;
+    top:10vh;
+	bottom: 0;
+	left: 85vw;
+	right: 0;
+}
+.profile-6{
+    height: 100px;
+    width: 100px;
+    top:76vh;
+	bottom: 0;
+	left: 50vw;
+	right: 0;
+}
+.profile-7{
+    height: 100px;
+    width: 100px;
+    top:80vh;
+	bottom: 0;
+	left: 88vw;
+	right: 0;
+}
+.profile-8{
+    height: 140px;
+    width: 140px;
+    top:70vh;
+	bottom: 0;
+	left: 90vw;
+	right: 0;
+}
+
+.background-container{
+    width:100%; 
+    height:100%;
+    position:fixed; 
+    left:0;
+    top:0;
+    overflow: hidden;
+}
+
+[class|="confetti"] {
+    position: absolute;
+  }
+  
+  .red {
+    background-color: #d13447;
+  }
+  
+  .yellow {
+    background-color: #ffbf00;
+  }
+  
+  .blue {
+    background-color: #263672;
+  }
+
+
+.detective{
+    z-index:777;
+}
+
+.img_detective{
+    width: 640px;;
+}
+
+
+
+.base-circle{
+    position: absolute;
+    top:0;
+	bottom: 0;
+	left: 0;
+	right: 0;
+    margin: auto;
+    border-radius: 50%;
+}
+
+
+.sphere-lg
+{
+    z-index: 2;
+}
+.sphere-med
+{
+	z-index: 3;
+}
+.sphere-small
+{
+    z-index: 4;
+}
+.page-header{
+    padding-top:6px;
+    padding-bottom: 6px;
+    z-index: 999;
+    position: fixed;
+    left: 0px;
+    top: 0px;
+    width: 100%;
+}
+
+.page-footer{
+    position: fixed;
+    z-index: 9999;
+    left: 0px;
+    padding-top: 12px !important;
+    padding-bottom: 8px !important;
+    bottom: 0px;
+    width: 100%;
+}
+
+.modal{
+    padding-left:0;
+}
+
+.modal-dialog {
+    max-width:800px;
+           }
+
+@media (min-width:576px){
+
+    .btn-start-quiz{
+        width: 100%;
+    }
+    
+    .detective{
+        margin-top: 5%;
+        margin-left:50%;
+    }
+    .sphere-lg
+    {
+       width: 1700px;
+       height: 1700px;
+       top:40%;
+       left:-20%;
+       right:20%;
+    }
+    .sphere-med
+    {
+        width: 1200px;
+       height: 1200px;
+       top:60%;
+       left:20%;
+    }
+    .sphere-small
+    {
+        width: 900px;
+        height: 900px;
+        top:60%;
+        left:40%;
+    }
+    
+
+}
+
+@media (min-width:768px){
+    .detective{
+        margin-top: 5%;
+        margin-left:40%;
+    }
+    .sphere-lg
+    {
+       width: 1400px;
+       height: 1400px;
+       top:60%;
+       right:0;
+    }
+    .sphere-med
+    {
+        width: 1000px;
+       height: 1000px;
+       top:60%;
+       left:10%;
+    }
+    .sphere-small
+    {
+        width: 500px;
+        height: 500px;
+        top:60%;
+        left:50%;
+    }
+  
+
+}
+
+@media (min-width:992px){
+    .detective{
+        margin-top: 25%;
+        margin-left:0;
+    }
+    .sphere-lg
+    {
+       width: 1400px;
+       height: 1400px;
+       top:0;
+       left:0;
+    }
+    .sphere-med
+    {
+        top:0;
+        width: 1000px;
+       height: 1000px;
+       left:20%;
+    }
+    .sphere-small
+    {
+        top:0;
+        width: 500px;
+        height: 500px;
+        left:20%;
+    }
+  
+}
+
+@media (min-width:1200px){
+    .detective{
+        margin-top: -15%;
+        margin-left:0%;
+    }
+    .sphere-lg
+    {
+       width: 100vw;
+       height: 100vw;
+       left:20%;
+    }
+    .sphere-med
+    {
+        width: 1000px;
+       height: 1000px;
+       left:40%;
+    }
+    .sphere-small
+    {
+        width: 500px;
+        height: 500px;
+        left:40%;
+    }
+    .btn-start-quiz{
+        width: 50%;
+    }
+
+  
+
+       .btn-quiz-answer {
+       min-width:130px;
+    }
+    
+}
+
+body{
+    overflow: hidden;
 }
